refactor: move Guestbook components into ui directory
